Description of data set + wrangling data
I tidied up the names and student number section so it was more legible. I wrote a short background on the data set and explained the variables in the dataset. I also added our question of interest and justification for choosing this question. I loaded in the data set and wrangled the data to fit proper  formatting.